Update dashboard.md - Danielle edit #1
Okay, there were definitely a few places here that tripped me up because I wasn't sure what was going on in the sentence without a visual reference and I could NOT get Dropbox to let me open the video files in Screenflow for some reason. (Going to restart my computer and see if that helps).
The following are key places I got lost:
-Line 9: "any new outs"? What does "outs" refer to?
-Line 29: This one confused me and I probably rephrased things slightly wrong here. Needs double checking.
-Line 37: Not sure if I phrased this line correctly either.
-Line 41: "not automatically secured" and "no security being applied" sort of sounds like the same thing, right? Also, not sure if that last sentence has the proper punctuation with ":". it sounded right when I put it there, but I'm not 100% sure.
-Line 45: Do the names of the admin types need special formatting? This is one instance where it would have been helpful to see the video side by side, but my computer is being weird about it today.
-Line 47: Punctuation okay?
-Line 57: This one really lost me and I haven't made changes to every sentence because I didn't know what changes to make. Does "overflow admin" need formatting of some kind? What is "Coran"? I think the last two sentences are mostly ok, though.

I think that covers most of the problem areas I found, but there might be some other items I overlooked while I was contemplating all of that.