Update fields.md - round #1
Just finished this edit of Fields, and I'm sure the biggest issue throughout might be proper formatting, like not knowing whether to include the "$" in a lot of places. Also, in general, I try to break up those big blocks of text into smaller, more manageable bites so it doesn't seem so overwhelming as a reader. Hope it has been ok so far.