Update post-action.md - Round #1
This is a small quarrel, but throughout the entire tutorial, you've used "Show" and "Detail" interchangeably when the page is called "Show". I've gone along with whatever you refer to it as in the moment generally, but the more I think about it, the more I believe that could be confusing for people who read one thing and hear another. You still mouse over what you're talking about so I'm sure they can figure it out, but I thought it was worth pointing out.